alertmanager/requirements.txt
bcrypt
pyyaml